# Java: fix for performance regression introduced in 3.8.0