--- a/java/core/src/main/java/com/google/protobuf/CodedOutputStream.java
+++ b/java/core/src/main/java/com/google/protobuf/CodedOutputStream.java
@@ -1348,35 +1348,14 @@ public abstract class CodedOutputStream extends ByteOutput {
       if (HAS_UNSAFE_ARRAY_OPERATIONS
           && !Android.isOnAndroidDevice()
           && spaceLeft() >= MAX_VARINT32_SIZE) {
-        int position = this.position;
-        try {
+        while (true) {
           if ((value & ~0x7F) == 0) {
             UnsafeUtil.putByte(buffer, position++, (byte) value);
             return;
+          } else {
+            UnsafeUtil.putByte(buffer, position++, (byte) ((value & 0x7F) | 0x80));
+            value >>>= 7;
           }
-          UnsafeUtil.putByte(buffer, position++, (byte) (value | 0x80));
-          value >>>= 7;
-          if ((value & ~0x7F) == 0) {
-            UnsafeUtil.putByte(buffer, position++, (byte) value);
-            return;
-          }
-          UnsafeUtil.putByte(buffer, position++, (byte) (value | 0x80));
-          value >>>= 7;
-          if ((value & ~0x7F) == 0) {
-            UnsafeUtil.putByte(buffer, position++, (byte) value);
-            return;
-          }
-          UnsafeUtil.putByte(buffer, position++, (byte) (value | 0x80));
-          value >>>= 7;
-          if ((value & ~0x7F) == 0) {
-            UnsafeUtil.putByte(buffer, position++, (byte) value);
-            return;
-          }
-          UnsafeUtil.putByte(buffer, position++, (byte) (value | 0x80));
-          value >>>= 7;
-          UnsafeUtil.putByte(buffer, position++, (byte) value);
-        } finally {
-          this.position = position;
         }
       } else {
         try {