This PR serves as a sample for the #6977 discussion. I don't have the toolchains setup anymore to run the benchmarks myself, but this minor change fixed it ~2 years ago.