@ennerf 2022-03-04
I ended up going through the benchmark setup, but I unfortunately couldn't reproduce the behavior I was seeing. There may have been other changes since then.

Here is a gist for a Java version of the Scala benchmark: [WriteUint32Bench](https://gist.github.com/ennerf/5a86471138ecedd7e04d7e7a87f4575a)