@ennerf 2022-03-04
reverting to the 3.7.0 version fixes it

```Java
/**
 * === 3.7.0
 * Benchmark                              Mode  Cnt   Score   Error   Units
 * WriteUint32Bench.testWriteInt32NoTag  thrpt   10  29,559 ± 0,499  ops/us
 * WriteUint32Bench.testWriteMessage     thrpt   10  71,854 ± 0,623  ops/us
 *
 * === 3.8.0
 * Benchmark                              Mode  Cnt   Score   Error   Units
 * WriteUint32Bench.testWriteInt32NoTag  thrpt   10   2,178 ± 0,016  ops/us
 * WriteUint32Bench.testWriteMessage     thrpt   10  33,218 ± 0,285  ops/us
 *
 * === 3.19.4
 * Benchmark                              Mode  Cnt   Score   Error   Units
 * WriteUint32Bench.testWriteInt32NoTag  thrpt   10   2,164 ± 0,027  ops/us
 * WriteUint32Bench.testWriteMessage     thrpt   10  32,970 ± 0,217  ops/us
 *
 * === 3.19.4 (reverted to 3.7.0 code)
 * Benchmark                              Mode  Cnt   Score   Error   Units
 * WriteUint32Bench.testWriteInt32NoTag  thrpt   10  29,605 ± 0,501  ops/us
 * WriteUint32Bench.testWriteMessage     thrpt   10  44,623 ± 1,239  ops/us
 *
 */
```