@ennerf 2022-03-04
Because it's faster. The Unsafe code was added as an optimization, but it only slows things down. The path that actually requires Unsafe (`UnsafeDirectNioEncoder`) is still using it.

I created separate commits so that you can cherry pick the version you want.