@ennerf 2022-03-04
I added the benchmark results to the previous comment. I only took out the "fast path", so the code is exactly the same as what was in the `else` statement before.