@elharo 2022-03-06
FYI, GWP can detect this function and shows it costing about 14,000 GCU so it is a plausible optimization target