@elharo 2022-03-06
Important to note: the optimization is on ArrayEncoder.writeUInt32NoTag rather than the method(s) a little higher in the call stack being benchmarked. That's not wrong, but it is important for following this proposal.