@elharo 2022-03-06
With change that eliminates the unsafe branch:

```
# Warmup Iteration   1: 7842926.924 ops/s
# Warmup Iteration   2: 12228825.174 ops/s
# Warmup Iteration   3: 12240813.322 ops/s
# Warmup Iteration   4: 12264911.947 ops/s
# Warmup Iteration   5: 12317264.162 ops/s
Iteration   1: 12227318.361 ops/s
Iteration   2: 12228675.497 ops/s
Iteration   3: 12227465.847 ops/s
Iteration   4: 12340207.375 ops/s
Iteration   5: 12325522.243 ops/s
```

Without change:

```
Duplicate Google agent found - skipping google3agentlib.
# Warmup Iteration   1: 2952518.421 ops/s
# Warmup Iteration   2: 2997562.659 ops/s
# Warmup Iteration   3: 3249345.823 ops/s
# Warmup Iteration   4: 3248366.059 ops/s
# Warmup Iteration   5: 3248051.117 ops/s
Iteration   1: 3248173.974 ops/s
Iteration   2: 3216751.094 ops/s
Iteration   3: 3237493.224 ops/s
Iteration   4: 3195381.124 ops/s
Iteration   5: 3225395.411 ops/s
```

As always with this sort of microbenchmark, the results can be highly dependent on the exact OS, hardware, and JDK. They could even reverse. However these results are from the OS  and JDK Google uses in production, so they are the ones Google cares most about. In that environment, this change is almost three times slower.