@ennerf 2022-03-06
> In that environment, this change is almost three times slower.

Removing unsafe branch: `12M ops/sec`
Current version: `3M ops/sec`

I'd interpret this as a 4x speedup. Am I misunderstanding something?