@elharo 2022-03-06
The only thing you're misunderstanding is that I can't read big numbers without commas. You're right. That is a 4x speedup which is worth pursuing here. Let me pass this around the team and see if anyone can poke holes in the benchmark, but otherwise this does look good. Thanks.