@elharo 2022-03-08
I went back about six years in the history and I think I found where the original "optimization" was introduced, cl/119195498 for the record. The benchmark results at the time showed a small improvement, less than 20%. The results were much more noticeable for some of the other encoders, but it's easy to believe that changes in the JVM or the native OS in the last six years could easily have turned the optimization negative along this code path.

What I plan to do is implement this change in google3 where we have much stronger tools for measuring performance and a very large sample of production code in which to test. If it works there it should save on the order of 1K GCUs, which is noticeable. If that works out, it will get synced back into open source.