@ennerf 2022-03-08
Sounds good. Feel free to close the PR if you don't need it anymore.

FWIW if you're looking for an easy performance win, you can change the generated merge statement to allow fallthrough. I've seen quite a few cases where this doubled the total decoding throughput, and I've yet to see anything that became slower.

```Java
int tag = input.readTag();
while(true) {
    switch(tag) {
        case FIELD_1_TAG: {
            // <parse field 1>
            if ((tag = input.readTag) != FIELD_2_TAG) {
                break;
            }
        }
        case FIELD_2_TAG: {
            // <parse field 2>
            if ((tag = input.readTag) != FIELD_3_TAG) {
                break;
            }
        }
        case FIELD_3_TAG: {
            // <parse field 3>
            // etc.
        }
    }
}
```